requirements_test.txt
# Test dependencies
pytest>=8.0.0
pytest-homeassistant-custom-component>=0.13.0
pytest-asyncio>=0.23.0
pycryptodome>=3.20.0
websockets>=15.0
aiohttp>=3.9.0
